refactor: add `onExit` to dashboard editor
#1316
