+++ dashboard/src/dashboard-editor/ui/header/index.tsx
@@ -2,6 +2,7 @@ import { observer } from 'mobx-react-lite';
 import { IDashbaordEditorHeaderMain, MainHeader } from './main-header';
 import { SubHeader } from './sub-header';
 
+
 interface IDashboardEditorHeader extends IDashbaordEditorHeaderMain {
   [key: string]: any;
 }
@@ -14,3 +15,5 @@ export const DashboardEditorHeader = observer((props: IDashboardEditorHeader) =>
     </>
   );
 });
+
+export type { OnExitParams, OnExitCallback } from './main-header';
