+++ dashboard/src/dashboard-editor/ui/header/main-header.tsx
@@ -1,68 +1,52 @@
 import { Button, Group, Header as MantineHeader, Text } from '@mantine/core';
-import { useModals } from '@mantine/modals';
-import { IconAlertTriangle, IconArrowLeft } from '@tabler/icons-react';
+import { IconArrowLeft } from '@tabler/icons-react';
 import { observer } from 'mobx-react-lite';
 import { ReactNode } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { useEditDashboardContext } from '~/contexts';
 import { HeaderMenu } from './header-menu';
 import { ISaveChangesOrMore, SaveChangesOrMore } from './save-changes-or-more';
 
+export type OnExitParams = { hasChanges: boolean; dashboardId: string };
+export type OnExitCallback = (params: OnExitParams) => void;
+
 export interface IDashbaordEditorHeaderMain extends ISaveChangesOrMore {
   headerSlot?: ReactNode;
+  onExit: OnExitCallback;
 }
 
-export const MainHeader = observer(({ saveDashboardChanges, headerSlot = null }: IDashbaordEditorHeaderMain) => {
-  const navigate = useNavigate();
-  const model = useEditDashboardContext();
+export const MainHeader = observer(
+  ({ saveDashboardChanges, onExit, headerSlot = null }: IDashbaordEditorHeaderMain) => {
+    const model = useEditDashboardContext();
 
-  const goBack = () => {
-    navigate(`/dashboard/${model.id}`);
-  };
-  const modals = useModals();
+    const hasChanges = model.content.changed;
+    const goBack = () => {
+      onExit({ hasChanges, dashboardId: model.id });
+    };
 
-  const goBackWithConfirmation = () => {
-    modals.openConfirmModal({
-      title: (
-        <Group position="left">
-          <IconAlertTriangle size={18} color="red" />
-          <Text>There are unsaved changes</Text>
+    return (
+      <MantineHeader height={60} px="md" py={0} sx={{ zIndex: 299 }}>
+        <Group position="apart" sx={{ height: 60, minWidth: '1000px', position: 'relative' }}>
+          <Group>
+            <Button
+              size="xs"
+              color={hasChanges ? 'red' : 'green'}
+              leftIcon={<IconArrowLeft size={20} />}
+              onClick={goBack}
+            >
+              <Group spacing={4}>
+                End Editing
+                <Text td="underline">{model.name}</Text>
+              </Group>
+            </Button>
+
+            <SaveChangesOrMore saveDashboardChanges={saveDashboardChanges} />
+          </Group>
+          <Group position="right" sx={{ flexGrow: 1 }}>
+            {headerSlot}
+            <HeaderMenu />
+          </Group>
         </Group>
-      ),
-      labels: { confirm: 'Discard', cancel: 'Cancel' },
-      confirmProps: { color: 'red' },
-      onCancel: () => console.log('Cancel'),
-      onConfirm: goBack,
-      zIndex: 320,
-      withCloseButton: false,
-    });
-  };
-
-  const hasChanges = model.content.changed;
-
-  return (
-    <MantineHeader height={60} px="md" py={0} sx={{ zIndex: 299 }}>
-      <Group position="apart" sx={{ height: 60, minWidth: '1000px', position: 'relative' }}>
-        <Group>
-          <Button
-            size="xs"
-            color={hasChanges ? 'red' : 'green'}
-            leftIcon={<IconArrowLeft size={20} />}
-            onClick={hasChanges ? goBackWithConfirmation : goBack}
-          >
-            <Group spacing={4}>
-              End Editing
-              <Text td="underline">{model.name}</Text>
-            </Group>
-          </Button>
-
-          <SaveChangesOrMore saveDashboardChanges={saveDashboardChanges} />
-        </Group>
-        <Group position="right" sx={{ flexGrow: 1 }}>
-          {headerSlot}
-          <HeaderMenu />
-        </Group>
-      </Group>
-    </MantineHeader>
-  );
-});
+      </MantineHeader>
+    );
+  },
+);
